fix invertCover logic for Lumi covers: align with device native convention
Lumi devices natively use 0=closed, 100=open. The default behavior (invertCover=false)
should now pass values as-is, while invertCover=true applies inversion.

+++ properties/tuya.cpp
@@ -311,7 +311,7 @@ void PropertiesTUYA::CoverMotor::update(quint8 dataPoint, const QVariant &data)
     {
         quint8 value = static_cast <quint8> (option("invertCover").toBool() ? data.toInt() : 100 - data.toInt());
         map.insert("cover", value ? "open" : "closed");
-        map.insert("position", static_cast <quint8> (value));
+        map.insert("position", value);
     }
 
     m_value = map.isEmpty() ? QVariant() : map;
